goal_tree/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
autopep8==1.7.0
cffi==1.14.5
cryptography==3.4.7
distlib==0.3.5
dj-database-url==1.0.0
Django==4.1
django-heroku==0.3.1
django-rest-knox==4.2.0
django-treebeard==4.5.1
djangorestframework==3.13.1
filelock==3.8.0
gunicorn==20.1.0
neat-python==0.92
numpy==1.23.2
pbr==5.10.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pygame==2.1.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
stevedore==4.0.0
toml==0.10.2
tzdata==2022.2
virtualenv==20.16.3
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
whitenoise==6.2.0
